now measuring correctly

+++ Lab 5/src/lab5.h
@@ -38,10 +38,10 @@ unsigned int getADCAtPin(unsigned char pin);
 float getVoltageAtPin(unsigned char pin);
 
 /* get period of signal based on comparitor zero cross input */
-double getHalfPeriod(void);
+float getHalfPeriod(void);
 
 /* get phase difference of the two waves */
-double getPeriodDiff(void);
+float getPeriodDiff(void);
 
 /* ===[MAX7219 CONTROL]=== */
 #define LED_CS P2_7
